--- a/test/modules/sungeun/init/printModuleInitOrder.good
+++ b/test/modules/sungeun/init/printModuleInitOrder.good
@@ -45,7 +45,6 @@ Initializing Modules:
     MemTracking
     ChapelUtil
     ChapelError
-    ChapelTaskData
     Builtins
     AlignedTSupport
     CPtr