--- a/test/modules/sungeun/init/printModuleInitOrder.na-none.good
+++ b/test/modules/sungeun/init/printModuleInitOrder.na-none.good
@@ -44,7 +44,6 @@ Initializing Modules:
     MemTracking
     ChapelUtil
     ChapelError
-    ChapelTaskData
     Builtins
     AlignedTSupport
     CPtr