Improve performance for single iteration coforalls like `coforall 1..1`.
You wouldn't write that code directly but it can come indirectly from
parallel iterators that are implemented with coforalls. Previously,
single iteration coforalls were pretty expensive so our internal
tuple/range/domain/array iterators all special case to avoid them.
This single task specialization improves execution performance, but
results in loop cloning which hurts compilation time. I'm hoping this PR
will allow us to remove that specialization. This results in a ~15x
speedup for serial single iteration coforalls and a ~90x speedup for
nested single iteration coforalls. There's still overhead for allocating
and manipulating the endcount, so it's not as cheap as the single task
specialization but the large sources of overhead are gone now.

This stops creating a task for single iteration coforalls, which gets
rid of task creation overhead and more importantly avoids spawning a new
task behind a task that's running on a different core and having to
wait. Without this optimization, removing the single task specialization
hurt ISx performance by 20% at 256 nodes. The implementation of this is
a little hacky because it's using task-local storage to store that the
next coforall task spawn should be serialized. We should probably do
this passing a flag but that required more compiler mods than I was
equipped to handle and I view this as a stopgop until we implement a
bulk/vector task spawn API (#9614).

This also avoids modifying the running task count for single task
coforalls where we'd end up adding/subtracting `0` anyways. The running
task counter is a shared atomic, so if you had an already parallel
section doing coforalls you'd have all tasks contending to add/sub `0`,
which hurt performance.

Here's a task creation microbenchmark run on chapcs (24-core server):

```chpl
use Time;

config const trials = 5_000_000;
var t: Timer; t.start();

for 1..trials do
  coforall 1..1 { }
writeln(t.elapsed()); t.clear();

coforall 1..here.maxTaskPar do
  for 1..trials do
    coforall 1..1 { }
writeln(t.elapsed()); t.clear();
```

| coforall type | before | now    |
| ------------- | ------ | ------ |
| serial        |  4.75s |  0.35s |
| nested        | 35.83s |  0.39s |

Closes https://github.com/Cray/chapel-private/issues/1306